# Use stderr instead of file