Hi,

Benchmark results (`wrk` results) are written in file, and this one is parsed as CSV.

To simplify process, it will be convenient to use stderr (memory) to write this CSV.

Regards,